# Optimized HashTable.index