@Ichoran 2016-11-17
This is fun--and I can't imagine the array lookup is actually faster than a proper LZCNT assembly instruction, so it indicates that the JIT compiler isn't actually emitting ideal assembly.  That suggests to me that we shouldn't bake in a table lookup because (1) if the JIT compiler improves the table will be slower than the alternatives, and (2) if the loop is less hot, the table lookup will be more expensive since it won't be hanging out in L1 cache the entire time.

But why not store tableLengthLeadingZeros and update it only when the table size changes (i.e. rarely)?  That should be faster yet.  Yes, it takes an extra 4 bytes of storage, but HashMap isn't exactly careful with how much it allocates as it is.