@paplorinc 2016-11-17
Valid concerns, @Ichoran!
I pushed a version with the built-in leading zeros counter, and will verify whether `HashMap`'s object layout permits an additional `byte` for free, and post the benchmark results here, so that we can take a informed decision :).
Thanks for all your reviews! :)